Substract background (from 2D data)

+++ SciAnalysis/XSAnalysis/Data.py
@@ -142,7 +142,7 @@ class Data2DScattering(Data2D):
     def load_tiff(self, infile):
         
         img = PIL.Image.open(infile).convert('I') # 'I' : 32-bit integer pixels
-        self.data = np.copy( np.asarray(img) )
+        self.data = (np.copy( np.asarray(img) )).astype(np.float)
         del img
